refactor puzzle grid into more files
